+++ src/Study/Config.hs
@@ -1,0 +1,19 @@
+-----------------------------------------------------------------------------
+--
+-- Module      :   Config
+
+-- the database and directory where the files are
+
+-----------------------------------------------------------------------------
+
+module Study.Config (filesUsed, dbPath) where
+
+databaseFolder = "/home/frank/afWorldDB/"
+
+databaseNmae = "r4cdb1" 
+
+dbPath = databaseFolder ++ databaseNmae ++ ".sqlite"
+
+filesUsed :: FilePath
+filesUsed =
+    "/home/frank/Desktop/buecher/worldFundamentals/WorldBankData"