+++ ChangeLog.md
@@ -3,6 +3,9 @@
 0.0.3 separate folder for each study. (baseTest, pt)
     produce md table (which can be sorted)
     add makedb to create a new db 
+    change structure of worldBank: keep zip files in book
+        automatically expand in a non-synced dir and load from there the cvs file
+        (not split for different topics!)
 
 a table 
 | Region | Population 2024 (M) | Surface 2023 (M km²) | Surface per capita (ha/person) |